backend/requirements.txt
# Flask Core
Flask==3.0.0
Werkzeug==3.0.1

# Base de Datos
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.7

# CORS
Flask-CORS==4.0.0

# Variables de Entorno
python-dotenv==1.0.0

# Autenticación (opcional)
Flask-JWT-Extended==4.5.3
bcrypt==4.1.2

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
catboost==1.2.2
joblib==1.3.2

# API y Validación
marshmallow==3.20.1
flask-marshmallow==0.15.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3

# Servidor de Producción
gunicorn==21.2.0
